@@ -4,28 +4,29 @@
   "private": true,
   "homepage": "",
   "dependencies": {
-    "@testing-library/jest-dom": "5.14.1",
-    "@testing-library/react": "12.0.0",
-    "@testing-library/user-event": "13.2.1",
-    "antd": "4.18.7",
-    "axios": "0.26.0",
-    "classnames": "2.3.1",
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "14.0.0",
+    "@testing-library/user-event": "14.4.3",
+    "@types/jest": "^29.5.1",
+    "antd": "5.4.6",
+    "axios": "1.4.0",
+    "classnames": "2.3.2",
     "formik": "2.2.9",
-    "gh-pages": "3.2.3",
-    "jest": "27.5.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-redux": "7.2.6",
-    "react-router-dom": "5.3.0",
-    "react-scripts": "5.0.0",
-    "react-test-renderer": "17.0.2",
-    "redux": "4.1.2",
-    "redux-form": "8.3.8",
-    "redux-thunk": "2.4.1",
-    "reselect": "4.1.5",
-    "typescript": "4.4.2",
-    "uuid": "8.3.2",
-    "web-vitals": "2.1.0"
+    "gh-pages": "5.0.0",
+    "jest": "29.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "8.0.5",
+    "react-router-dom": "^6.11.1",
+    "react-scripts": "^5.0.1",
+    "react-test-renderer": "18.2.0",
+    "redux": "4.2.1",
+    "redux-form": "8.3.10",
+    "redux-thunk": "2.4.2",
+    "reselect": "4.1.8",
+    "typescript": "5.0.4",
+    "uuid": "9.0.0",
+    "web-vitals": "3.3.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -54,13 +55,13 @@
     ]
   },
   "devDependencies": {
-    "@types/node": "16.7.13",
-    "@types/react": "17.0.20",
-    "@types/react-dom": "17.0.9",
-    "@types/react-redux": "7.1.22",
-    "@types/react-router-dom": "5.3.3",
-    "@types/react-test-renderer": "17.0.1",
-    "@types/redux-form": "8.3.3",
-    "@types/uuid": "8.3.4"
+    "@types/node": "^18.16.3",
+    "@types/react": "^18.2.2",
+    "@types/react-dom": "^18.2.3",
+    "@types/react-redux": "^7.1.25",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-test-renderer": "^18.0.0",
+    "@types/redux-form": "^8.3.5",
+    "@types/uuid": "^9.0.1"
   }
 }
